+++ components/usb/usb_stream/CHANGELOG.md
@@ -1,5 +1,48 @@
 # ChangeLog
 
+## v1.0.0 - 2023-05-10
+
+### Release v1.0.0
+
+### Enhancements:
+
+* USB:
+  * Refactor USB state machine, handle error and hot-plug more robustly.
+  * Support user submit control transfer at any time (Better to support more control requests in future).
+  * Support self-define initial state through `flags`, eg `FLAG_UAC_SPK_SUSPEND_AFTER_START` to make speaker suspend as startup state.
+  * Support Stream Quick Start Mode (Skip the enum process to save at least 50 ms).
+
+* UVC:
+  * Support get camera resolution list and current resolution.
+  * Support dynamically config UVC frame rate and resolution.
+
+* UAC:
+  * Support config UAC `*_ch_num`, `*_samples_frequence`, `*_bit_resolution` to `*_ANY` to adapter any available device.
+  * Support get audio frame rate and resolution list.
+  * Support dynamically config audio channel number, rate, and bits.
+
+* Test:
+  * Add hardware test for usb_stream component.
+
+### Bug Fixes:
+  * Fix some hot-plug issues
+  * Fix UAC control interface can not be 0
+  * Fix UAC stereo headset volume control failed
+
+### Breaking Changes:
+
+* UAC:
+  * Remove the configuration `mic_min_bytes` from `uac_config_t`, use KConfig option `CONFIG_UAC_MIC_CB_MIN_MS_DEFAULT` instead. The default value is 20 ms, which means the mic callback will be called every 20 ms.
+  * Add the configuration `flags` to `uac_config_t` and `uvc_config_t`, users can use it to define the initial state of the device. The default value is 0.
+  * Add the configuration `mic_ch_num` and `spk_ch_num` to `uac_config_t`, users can specify the number of the microphone and speaker channels. The default value is 0, which means support any number of channels.
+
+## v0.4.13 - 2023-04-13
+
+### Enhancements
+
+* test_apps:
+  * Update test_apps project
+
 ## v0.4.12 - 2023-04-12
 
 ### Bug Fixes: